move PropertiesProvider to org.cfg4j.source.context.propertiesprovider package
