+++ cfg4j-core/src/test/resources/org/cfg4j/source/propertiesprovider/PropertyBasedPropertiesProviderTest_shouldThrowOnMalformedFile.properties
