+++ cfg4j-core/src/test/resources/org/cfg4j/source/propertiesprovider/YamlBasedPropertiesProviderTest_shouldSupportReferences.yaml
