+++ cfg4j-git/src/main/java/org/cfg4j/source/git/GitConfigurationSource.java
@@ -22,9 +22,9 @@ import org.cfg4j.source.SourceCommunicationException;
 import org.cfg4j.source.context.environment.Environment;
 import org.cfg4j.source.context.environment.MissingEnvironmentException;
 import org.cfg4j.source.context.filesprovider.ConfigFilesProvider;
+import org.cfg4j.source.context.propertiesprovider.PropertiesProvider;
+import org.cfg4j.source.context.propertiesprovider.PropertiesProviderSelector;
 import org.cfg4j.utils.FileUtils;
-import org.cfg4j.utils.PropertiesProvider;
-import org.cfg4j.utils.PropertiesProviderSelector;
 import org.eclipse.jgit.api.CheckoutCommand;
 import org.eclipse.jgit.api.CreateBranchCommand;
 import org.eclipse.jgit.api.Git;
