+++ cfg4j-git/src/main/java/org/cfg4j/source/git/GitConfigurationSourceBuilder.java
@@ -17,9 +17,9 @@ package org.cfg4j.source.git;
 
 import org.cfg4j.source.context.filesprovider.ConfigFilesProvider;
 import org.cfg4j.source.context.filesprovider.DefaultConfigFilesProvider;
-import org.cfg4j.utils.PropertiesProviderSelector;
-import org.cfg4j.utils.PropertyBasedPropertiesProvider;
-import org.cfg4j.utils.YamlBasedPropertiesProvider;
+import org.cfg4j.source.context.propertiesprovider.PropertiesProviderSelector;
+import org.cfg4j.source.context.propertiesprovider.PropertyBasedPropertiesProvider;
+import org.cfg4j.source.context.propertiesprovider.YamlBasedPropertiesProvider;
 
 /**
  * Builder for {@link GitConfigurationSource}.
